Progress Bar calculations added, visibility crash when another fragment is open not fixed

+++ app/src/main/java/ee364h/wearablehealthtracker/GraphType.java
@@ -24,7 +24,7 @@ public enum GraphType {
     },
     TEMPERATURE("Temperature",3,"Temperature","degrees Celsius"){
 	    public Number getValueFromStringArray(String [] values){
-	        return Float.valueOf(values[this.getValueLocation()]);
+	        return Double.valueOf(values[this.getValueLocation()]);
     	}
     };
